updated statistic printing for new rules and conclusions; improved formatting

+++ elk-reasoner/src/main/java/org/semanticweb/elk/reasoner/saturation/conclusions/visitors/ConclusionCounter.java
@@ -143,7 +143,7 @@ public class ConclusionCounter {
 		return countSubContextInitializations;
 	}
 
-	public long getTotalCount() {
+	public int getTotalCount() {
 		return countDecomposedSubsumers + countComposedSubsumers
 				+ countBackwardLinks + countForwardLinks + countContradictions
 				+ countPropagations + countDisjointSubsumers
